Moved chart colors into their own data file
